// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigInt: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};
// ...
export type ActivityEventPage = {
  items: Array<ActivityEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AddNftTierEventPage = {
  items: Array<AddNftTierEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AddToBalanceEventPage = {
  items: Array<AddToBalanceEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AutoIssueEventPage = {
  items: Array<AutoIssueEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BorrowLoanEventPage = {
  items: Array<BorrowLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BridgeClaimEventPage = {
  items: Array<BridgeClaimEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BurnEventPage = {
  items: Array<BurnEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BuybackPoolEventPage = {
  items: Array<BuybackPoolEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type CashOutTokensEventPage = {
  items: Array<CashOutTokensEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type DeployErc20EventPage = {
  items: Array<DeployErc20Event>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type LiquidateLoanEventPage = {
  items: Array<LiquidateLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ManualBurnEventPage = {
  items: Array<ManualBurnEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ManualMintTokensEventPage = {
  items: Array<ManualMintTokensEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type Project = {
  activityEvents: Maybe<ActivityEventPage>;
  addNftTierEvents: Maybe<AddNftTierEventPage>;
  addToBalanceEvents: Maybe<AddToBalanceEventPage>;
  autoIssueEvents: Maybe<AutoIssueEventPage>;
  balance: Scalars['BigInt']['output'];
  borrowLoanEvents: Maybe<BorrowLoanEventPage>;
  bridgeClaimEvents: Maybe<BridgeClaimEventPage>;
  burnEvents: Maybe<BurnEventPage>;
  buybackPoolEvents: Maybe<BuybackPoolEventPage>;
  cashOutTokensEvents: Maybe<CashOutTokensEventPage>;
  chainId: Scalars['Int']['output'];
  contributorsCount: Scalars['Int']['output'];
  coverImageUri: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Int']['output'];
  createdWithinTrendingWindow: Maybe<Scalars['Boolean']['output']>;
  creator: Scalars['String']['output'];
  currency: Maybe<Scalars['BigInt']['output']>;
  decimals: Maybe<Scalars['Int']['output']>;
  deployErc20Events: Maybe<DeployErc20EventPage>;
  deployer: Scalars['String']['output'];
  description: Maybe<Scalars['String']['output']>;
  discord: Maybe<Scalars['String']['output']>;
  domain: Maybe<Scalars['String']['output']>;
  farcaster: Maybe<Scalars['String']['output']>;
  handle: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  infoUri: Maybe<Scalars['String']['output']>;
  isRevnet: Maybe<Scalars['Boolean']['output']>;
  liquidateLoanEvents: Maybe<LiquidateLoanEventPage>;
  logoUri: Maybe<Scalars['String']['output']>;
  manualBurnEvents: Maybe<ManualBurnEventPage>;
  manualMintTokensEvents: Maybe<ManualMintTokensEventPage>;
  metadata: Maybe<Scalars['JSON']['output']>;
  metadataUri: Maybe<Scalars['String']['output']>;
  mintNftEvents: Maybe<MintNftEventPage>;
  mintTokensEvents: Maybe<MintTokensEventPage>;
  name: Maybe<Scalars['String']['output']>;
  nftHooks: Maybe<NftHookPage>;
  nfts: Maybe<NftPage>;
  nftsMintedCount: Scalars['Int']['output'];
  operatorPermissionsSetEvents: Maybe<OperatorPermissionsSetEventPage>;
  owner: Scalars['String']['output'];
  participants: Maybe<ParticipantPage>;
  payDisclosure: Maybe<Scalars['String']['output']>;
  payEvents: Maybe<PayEventPage>;
  paymentsCount: Scalars['Int']['output'];
  permissionHolders: Maybe<PermissionHolderPage>;
  projectCreateEvents: Maybe<ProjectCreateEventPage>;
  projectId: Scalars['Int']['output'];
  projectMoments: Maybe<ProjectMomentPage>;
  projectPayers: Maybe<ProjectPayerPage>;
  projectTagline: Maybe<Scalars['String']['output']>;
  projectTransferEvents: Maybe<ProjectTransferEventPage>;
  reallocateLoanEvents: Maybe<ReallocateLoanEventPage>;
  redeemCount: Scalars['Int']['output'];
  redeemVolume: Scalars['BigInt']['output'];
  redeemVolumeUsd: Scalars['BigInt']['output'];
  removeNftTierEvents: Maybe<RemoveNftTierEventPage>;
  repayLoanEvents: Maybe<RepayLoanEventPage>;
  reservedTokenSupply: Scalars['BigInt']['output'];
  rulesetQueuedEvents: Maybe<RulesetQueuedEventPage>;
  sendPayoutToSplitEvents: Maybe<SendPayoutToSplitEventPage>;
  sendPayoutsEvents: Maybe<SendPayoutsEventPage>;
  sendReservedTokensToSplitEvents: Maybe<SendReservedTokensToSplitEventPage>;
  sendReservedTokensToSplitsEvents: Maybe<SendReservedTokensToSplitsEventPage>;
  setUriEvents: Maybe<SetUriEventPage>;
  storeAutoIssuanceAmountEvent: Maybe<StoreAutoIssuanceAmountEventPage>;
  suckerGroup: Maybe<SuckerGroup>;
  suckerGroupId: Scalars['String']['output'];
  swapEvents: Maybe<SwapEventPage>;
  tags: Maybe<Array<Scalars['String']['output']>>;
  telegram: Maybe<Scalars['String']['output']>;
  token: Maybe<Scalars['String']['output']>;
  tokenSupply: Scalars['BigInt']['output'];
  tokenSymbol: Maybe<Scalars['String']['output']>;
  tokens: Maybe<Array<Scalars['String']['output']>>;
  trendingPaymentsCount: Scalars['Int']['output'];
  trendingScore: Scalars['BigInt']['output'];
  trendingVolume: Scalars['BigInt']['output'];
  trendingVolumeUsd: Scalars['BigInt']['output'];
  twitter: Maybe<Scalars['String']['output']>;
  useAllowanceEvents: Maybe<UseAllowanceEventPage>;
  version: Scalars['Int']['output'];
  volume: Scalars['BigInt']['output'];
  volumeUsd: Scalars['BigInt']['output'];
};
// ...
export type SuckerGroup = {
  addresses: Array<Scalars['String']['output']>;
  balance: Scalars['BigInt']['output'];
  contributorsCount: Scalars['Int']['output'];
  createdAt: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  nftsMintedCount: Scalars['Int']['output'];
  paymentsCount: Scalars['Int']['output'];
  projects: Maybe<ProjectPage>;
  redeemCount: Scalars['Int']['output'];
  redeemVolume: Scalars['BigInt']['output'];
  redeemVolumeUsd: Scalars['BigInt']['output'];
  reservedTokenSupply: Scalars['BigInt']['output'];
  suckerTransactions: Maybe<SuckerTransactionPage>;
  tokenSupply: Scalars['BigInt']['output'];
  trendingPaymentsCount: Scalars['Int']['output'];
  trendingScore: Scalars['BigInt']['output'];
  trendingVolume: Scalars['BigInt']['output'];
  version: Scalars['Int']['output'];
  volume: Scalars['BigInt']['output'];
  volumeUsd: Scalars['BigInt']['output'];
};
// ...
export type ProjectQueryVariables = Exact<{
  projectId: Scalars['Float']['input'];
  chainId: Scalars['Float']['input'];
  version: Scalars['Float']['input'];
}>;


export type ProjectQuery = { project: { projectId: number, metadataUri: string | null, handle: string | null, createdAt: number, suckerGroupId: string, logoUri: string | null, name: string | null, version: number, token: string | null, decimals: number | null, currency: any | null, tokenSymbol: string | null } | null };
// ...
export type SuckerGroupQueryVariables = Exact<{
  id: Scalars['String']['input'];
}>;


export type SuckerGroupQuery = { suckerGroup: { id: string, tokenSupply: any, projects: { items: Array<{ balance: any, chainId: number, currency: any | null, decimals: number | null, projectId: number, token: string | null, tokenSupply: any, tokenSymbol: string | null, version: number, suckerGroupId: string }> } | null } | null };
// ...
export const ProjectDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"Project"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"projectId"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"chainId"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"version"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Float"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"project"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"projectId"},"value":{"kind":"Variable","name":{"kind":"Name","value":"projectId"}}},{"kind":"Argument","name":{"kind":"Name","value":"chainId"},"value":{"kind":"Variable","name":{"kind":"Name","value":"chainId"}}},{"kind":"Argument","name":{"kind":"Name","value":"version"},"value":{"kind":"Variable","name":{"kind":"Name","value":"version"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"projectId"}},{"kind":"Field","name":{"kind":"Name","value":"metadataUri"}},{"kind":"Field","name":{"kind":"Name","value":"handle"}},{"kind":"Field","name":{"kind":"Name","value":"createdAt"}},{"kind":"Field","name":{"kind":"Name","value":"suckerGroupId"}},{"kind":"Field","name":{"kind":"Name","value":"logoUri"}},{"kind":"Field","name":{"kind":"Name","value":"name"}},{"kind":"Field","name":{"kind":"Name","value":"version"}},{"kind":"Field","name":{"kind":"Name","value":"token"}},{"kind":"Field","name":{"kind":"Name","value":"decimals"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSymbol"}}]}}]}}]} as unknown as DocumentNode<ProjectQuery, ProjectQueryVariables>;
// ...
export const SuckerGroupDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"SuckerGroup"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"suckerGroup"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSupply"}},{"kind":"Field","name":{"kind":"Name","value":"projects"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"items"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"balance"}},{"kind":"Field","name":{"kind":"Name","value":"chainId"}},{"kind":"Field","name":{"kind":"Name","value":"currency"}},{"kind":"Field","name":{"kind":"Name","value":"decimals"}},{"kind":"Field","name":{"kind":"Name","value":"projectId"}},{"kind":"Field","name":{"kind":"Name","value":"token"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSupply"}},{"kind":"Field","name":{"kind":"Name","value":"tokenSymbol"}},{"kind":"Field","name":{"kind":"Name","value":"version"}},{"kind":"Field","name":{"kind":"Name","value":"suckerGroupId"}}]}}]}}]}}]}}]} as unknown as DocumentNode<SuckerGroupQuery, SuckerGroupQueryVariables>;